requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
cloudinary==1.41.0
Flask-Limiter==3.8.0
urllib3==2.2.3